changed modules logic
