+++ index.html
@@ -7,10 +7,8 @@
   <link href="css/style.css" rel="stylesheet">
 </head>
 <body>
-
 <main class="central">
 </main>
-<script src="./js/main.js"></script>
-<script src="./js/rulesLayout.js"></script>
+<script src="./js/introScreen.js"></script>
 </body>
 </html>
