+++ js/game2Screen.js
@@ -1,4 +1,8 @@
-const game1Layout = `<div id="game-1-div" class="game-1 central__content">
+import showScreen from './showScreen';
+import showGame3Screen from './game3Screen';
+import checkToGreetingScreen from './checkToGreetingScreen';
+
+const game2Layout = `<div id="game-2-div" class="game-2 central__content">
     <header class="header">
       <div class="header__back">
         <button class="back">
@@ -14,27 +18,16 @@ const game1Layout = `<div id="game-1-div" class="game-1 central__content">
       </div>
     </header>
     <div class="game">
-      <p class="game__task">Угадайте для каждого изображения фото или рисунок?</p>
-      <form class="game__content">
+      <p class="game__task">Угадай, фото или рисунок?</p>
+      <form class="game__content  game__content--wide">
         <div class="game__option">
-          <img src="http://placehold.it/468x458" alt="Option 1" width="468" height="458">
-          <label class="game__answer game__answer--photo">
+          <img src="http://placehold.it/705x455" alt="Option 1" width="705" height="455">
+          <label class="game__answer  game__answer--photo">
             <input name="question1" type="radio" value="photo">
             <span>Фото</span>
           </label>
-          <label class="game__answer game__answer--paint">
+          <label class="game__answer  game__answer--wide  game__answer--paint">
             <input name="question1" type="radio" value="paint">
-            <span>Рисунок</span>
-          </label>
-        </div>
-        <div class="game__option">
-          <img src="http://placehold.it/468x458" alt="Option 2" width="468" height="458">
-          <label class="game__answer  game__answer--photo">
-            <input name="question2" type="radio" value="photo">
-            <span>Фото</span>
-          </label>
-          <label class="game__answer  game__answer--paint">
-            <input name="question2" type="radio" value="paint">
             <span>Рисунок</span>
           </label>
         </div>
@@ -45,11 +38,11 @@ const game1Layout = `<div id="game-1-div" class="game-1 central__content">
           <li class="stats__result stats__result--slow"></li>
           <li class="stats__result stats__result--fast"></li>
           <li class="stats__result stats__result--correct"></li>
+          <li class="stats__result stats__result--wrong"></li>
           <li class="stats__result stats__result--unknown"></li>
+          <li class="stats__result stats__result--slow"></li>
           <li class="stats__result stats__result--unknown"></li>
-          <li class="stats__result stats__result--unknown"></li>
-          <li class="stats__result stats__result--unknown"></li>
-          <li class="stats__result stats__result--unknown"></li>
+          <li class="stats__result stats__result--fast"></li>
           <li class="stats__result stats__result--unknown"></li>
         </ul>
       </div>
@@ -66,4 +59,18 @@ const game1Layout = `<div id="game-1-div" class="game-1 central__content">
     </footer>
   </div>`;
 
-export default game1Layout;
+
+const showGame2Screen = function () {
+  showScreen(game2Layout, checkToGame3Screen);
+};
+
+const checkToGame3Screen = function () {
+  checkToGreetingScreen();
+  const gameAnswers = document.querySelectorAll(`.game__answer`);
+
+  gameAnswers.forEach(function (btn) {
+    btn.addEventListener(`click`, showGame3Screen);
+  });
+};
+
+export default showGame2Screen;
