+++ js/game3Screen.js
@@ -1,4 +1,8 @@
-const game2Layout = `<div id="game-2-div" class="game-2 central__content">
+import showScreen from './showScreen';
+import showStatsScreen from './statsScreen';
+import checkToGreetingScreen from './checkToGreetingScreen';
+
+const game3Layout = `<div id="game-3-div" class="game-3 central__content">
     <header class="header">
       <div class="header__back">
         <button class="back">
@@ -14,18 +18,16 @@ const game2Layout = `<div id="game-2-div" class="game-2 central__content">
       </div>
     </header>
     <div class="game">
-      <p class="game__task">Угадай, фото или рисунок?</p>
-      <form class="game__content  game__content--wide">
+      <p class="game__task">Найдите рисунок среди изображений</p>
+      <form class="game__content  game__content--triple">
         <div class="game__option">
-          <img src="http://placehold.it/705x455" alt="Option 1" width="705" height="455">
-          <label class="game__answer  game__answer--photo">
-            <input name="question1" type="radio" value="photo">
-            <span>Фото</span>
-          </label>
-          <label class="game__answer  game__answer--wide  game__answer--paint">
-            <input name="question1" type="radio" value="paint">
-            <span>Рисунок</span>
-          </label>
+          <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
+        </div>
+        <div class="game__option  game__option--selected">
+          <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
+        </div>
+        <div class="game__option">
+          <img src="http://placehold.it/304x455" alt="Option 1" width="304" height="455">
         </div>
       </form>
       <div class="stats">
@@ -55,4 +57,17 @@ const game2Layout = `<div id="game-2-div" class="game-2 central__content">
     </footer>
   </div>`;
 
-export default game2Layout;
+const showGame3Screen = function () {
+  showScreen(game3Layout, checkToStatsScreen);
+};
+
+const checkToStatsScreen = function () {
+  checkToGreetingScreen();
+  const gameOptions = document.querySelectorAll(`.game__option`);
+
+  gameOptions.forEach(function (btn) {
+    btn.addEventListener(`click`, showStatsScreen);
+  });
+};
+
+export default showGame3Screen;
