+++ js/greetingScreen.js
@@ -1,3 +1,6 @@
+import showScreen from './showScreen';
+import showRulesScreen from './rulesScreen';
+
 const greetingLayout = `
 <div id="greeting-div" class="greeting central__content">
   <div class="greeting central--blur">
@@ -26,6 +29,13 @@ const greetingLayout = `
 </div>
 `;
 
+const showGreetingScreen = function () {
+  showScreen(greetingLayout, checkToRulesScreen);
+};
 
-export default greetingLayout;
+const checkToRulesScreen = function () {
+  const greetingContinue = document.querySelector(`.greeting__continue`);
+  greetingContinue.addEventListener(`click`, showRulesScreen);
+};
 
+export default showGreetingScreen;
